Custom Exception Package and updated Controller

+++ src/main/java/com/note/Notes/Service/UserServiceImplementation.java
@@ -1,6 +1,9 @@
 package com.note.Notes.Service;
 
 import com.note.Notes.Entity.User;
+import com.note.Notes.Exceptions.DuplicateUser;
+import com.note.Notes.Exceptions.NotesNotFoundException;
+import com.note.Notes.Exceptions.UserNotFoundException;
 import com.note.Notes.Repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,29 +16,44 @@ public class UserServiceImplementation implements UserServiceInterface{
     private UserRepository userRepository;
 
     @Override
-    public void createUser(User user) {
-        userRepository.save(user);
+    public String createUser(User user) {
+       Optional<User>userDb=userRepository.findByUserName(user.getUserName());
+        if(!userDb.isPresent())
+        {
+            userRepository.save(user);
+            return ("User is Created");
+        }else
+           throw new DuplicateUser("User is already exist :"+user.getUserName());
+
     }
 
     @Override
-    public User getuserInfo(long id) {
-        Optional<User>user1=userRepository.findById(id);
+    public User getuserInfo(String userName) {
+        Optional<User>user1=userRepository.findByUserName(userName);
         if(user1.isPresent())
         {
            User userDp=user1.get();
            return userDp;
         }
-        return null;
+        throw new UserNotFoundException("User does Found with having UserName :"+userName);
     }
 
     @Override
-    public void deleteById(long id) {
-       userRepository.deleteById(id);
+    public String deleteById(long id) {
+        Optional<User>user1=userRepository.findById(id);
+        if(user1.isPresent())
+        {
+            userRepository.deleteById(id);
+            return "User is Deleted";
+        }
+        throw new UserNotFoundException("User does Found with having id "+id);
     }
 
+
+
     @Override
-    public User updateUser(long id, User user) {
-       Optional<User>userDb=userRepository.findById(id);
+    public User updateUser(String userName, User user) {
+       Optional<User>userDb=userRepository.findByUserName(userName);
        if(userDb.isPresent())
        {
            User updatedUser=userDb.get();
@@ -53,6 +71,7 @@ public class UserServiceImplementation implements UserServiceInterface{
            userRepository.save(updatedUser);
            return updatedUser;
        }
-       return null;
+     throw new UserNotFoundException("User does not found with UserName :" +userName);
     }
 }
+
